requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
bcrypt==4.1.2
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML and Transformers
torch==2.1.1
transformers==4.36.0
tokenizers==0.15.0

# HTTP Client for Testing
requests==2.31.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality and Linting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0
